--- MANIFEST.in
@@ -1,12 +1,0 @@
-include LICENSE
-include README.md
-
-include setup.py
-include setup.cfg
-include pyproject.toml
-
-include package.json
-include higlass-widget.json
-
-graft higlass_widget/labextension
-graft higlass_widget/nbextension
