--- higlass-widget.json
@@ -1,5 +1,0 @@
-{
-  "load_extensions": {
-    "higlass-widget/extension": true
-  }
-}
