--- higlass_widget/nbextension/extension.js
@@ -1,15 +1,0 @@
-// Entry point for the notebook bundle containing custom model definitions.
-define(function() {
-	"use strict";
-	window.requirejs?.config({
-		map: {
-			'*': {
-				'higlass-widget': 'nbextensions/higlass-widget/index',
-			},
-		}
-	})
-	// Export the required load_ipython_extension function
-	return {
-		load_ipython_extension: function() {}
-	};
-});
