+++ higlass_widget/widget.py
@@ -1,27 +1,20 @@
+from __future__ import annotations
 import json
-from typing import Any, Dict
+import pathlib
 
-import ipywidgets
-import traitlets.traitlets as t
-
-from ._version import __version__
+import anywidget
+import traitlets as t
 
 
-class HiGlassWidget(ipywidgets.DOMWidget):
-    _model_name = t.Unicode("HiGlassModel").tag(sync=True)
-    _model_module = t.Unicode("higlass-widget").tag(sync=True)
-    _model_module_version = t.Unicode(__version__).tag(sync=True)
-
-    _view_name = t.Unicode("HiGlassView").tag(sync=True)
-    _view_module = t.Unicode("higlass-widget").tag(sync=True)
-    _view_module_version = t.Unicode(__version__).tag(sync=True)
-
+class HiGlassWidget(anywidget.AnyWidget):
+    _esm = (pathlib.Path(__file__).parent / "widget.js").read_text()
+    _css = "https://esm.sh/higlass@1.11/dist/hglib.css"
     _viewconf = t.Unicode("null").tag(sync=True)
 
     # readonly properties
     location = t.List(t.Union([t.Float(), t.Tuple()]), read_only=True).tag(sync=True)
 
-    def __init__(self, viewconf: Dict[str, Any], **kwargs):
+    def __init__(self, viewconf: dict, **kwargs):
         super().__init__(_viewconf=json.dumps(viewconf), **kwargs)
 
     def reload(self, *items):
@@ -33,8 +26,8 @@ class HiGlassWidget(ipywidgets.DOMWidget):
         view_id: str,
         start1: int,
         end1: int,
-        start2: int = None,
-        end2: int = None,
+        start2: int | None = None,
+        end2: int | None = None,
         animate_time: int = 500,
     ):
         msg = json.dumps(["zoomTo", view_id, start1, end1, start2, end2, animate_time])
