--- package.json
@@ -1,42 +1,0 @@
-{
-  "name": "higlass-widget",
-  "version": "0.0.7",
-  "main": "src/index.js",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "run-p build:nb build:lab",
-    "build:nb": "npm run bundle -- --outfile=higlass_widget/nbextension/index.js",
-    "build:lab": "jupyter labextension build .",
-    "bundle": "esbuild --format=esm --bundle --minify --loader:.css=text src/index.js",
-    "clean": "rimraf higlass_widget/nbextension/index.js higlass_widget/labextension/",
-    "prepack": "rimraf dist && npm run bundle -- --outfile=dist/index.js"
-  },
-  "keywords": [],
-  "jupyterlab": {
-    "extension": "src/plugin",
-    "outputDir": "higlass_widget/labextension",
-    "sharedPackages": {
-      "@jupyter-widgets/base": {
-        "bundled": false,
-        "singleton": true
-      }
-    }
-  },
-  "devDependencies": {
-    "@jupyter-widgets/base": "^4.0.0",
-    "@jupyterlab/builder": "^3.2.9",
-    "css-loader": "^6.6.0",
-    "esbuild": "^0.14.21",
-    "npm-run-all": "^4.1.5",
-    "rimraf": "^3.0.2",
-    "style-loader": "^3.3.1"
-  },
-  "dependencies": {
-    "higlass": "^1.11.9",
-    "pixi.js": "^6.2.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
-  }
-}
