+++ pyproject.toml
@@ -1,3 +1,33 @@
 [build-system]
-requires = ["jupyter_packaging>=0.10,<2", "jupyterlab==3.*"]
-build-backend = "setuptools.build_meta"
+requires = ["hatchling", "hatch-vcs"]
+build-backend = "hatchling.build"
+
+[project]
+name = "higlass-widget"
+description = "A Jupyter Widget for HiGlass"
+authors = [
+    { name = "Trevor Manz", email = "trevor.j.manz@gmail.com" },
+]
+license = { text = "MIT" }
+dynamic = ["version"]
+readme = "README.md"
+requires-python = ">=3.8"
+dependencies = [
+    "anywidget",
+]
+
+[project.optional-dependencies]
+dev = [
+    "black[jupyter]",
+    "ruff",
+    "jupyterlab"
+]
+
+[project.urls]
+homepage = "https://github.com/higlass/higlass-widget"
+
+[tool.hatch.version]
+source = "vcs"
+
+[tool.hatch.envs.default]
+features = ["dev"]
