--- setup.cfg
@@ -1,18 +1,0 @@
-[metadata]
-name = higlass-widget
-author = Trevor Manz
-author_email = trevor.j.manz@gmail.com
-license = MIT
-description = A Jupyter Widget for HiGlass
-url = https://github.com/higlass/higlass-widget
-long_description = file: README.md
-long_description_content_type = text/markdown
-
-[options]
-include_package_data = True
-packages = find:
-install_requires = ipywidgets
-python_requires = >=3.7
-
-[options.package_data]
-* = nbextension/*.js, labextension/**/*, package.json
